+++ src/App.jsx
@@ -3,14 +3,16 @@ import { globalStyle } from '@/styles/globals';
 import { Global, ThemeProvider } from '@emotion/react';
 import { Layout } from '@components/global/Layout';
 import Home from '@pages/home/Home';
-import FrontendCur from './pages/home/FrontendCur';
-import BackendCur from './pages/home/BackendCur';
+import FrontendCur from '@pages/home/FrontendCur';
+import BackendCur from '@pages/home/BackendCur';
+import FAQ from '@pages/home/FAQ';
 
 function App() {
   return (
     <Layout>
       <Global styles={globalStyle} />
-      <Home />
+      {/* <Home /> */}
+      <FAQ />
     </Layout>
   );
 }
