+++ src/pages/home/FAQ.jsx
@@ -1,0 +1,12 @@
+import { colors } from '@/styles/colors';
+import FAQList from '@components/home/FAQList';
+
+const FAQ = () => {
+  return (
+    <div>
+      <FAQList listNum={13} FAQcolor={colors.black} />
+    </div>
+  );
+};
+
+export default FAQ;
